pyproject.toml
[tool.poetry]
name = "practiceapi"
version = "0.1.0"
description = ""
authors = ["dianasokolovaa <sokolovad46@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
psycopg2 = "^2.9.10"
asyncio = "^3.4.3"
fastapi = "^0.115.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
